+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 1.3
 - Hide QR option on web
+- Fix removal of old detected shortcuts
+- Use larger Capacitor icon
 
 ## 1.2
 - Add support for appflow.app.com QR codes